feat: add rne
